requirements.txt
certifi==2022.5.18.1
charset-normalizer==2.0.12
idna==3.3
pyaes==1.6.1
pyasn1==0.4.8
python-dotenv==0.20.0
requests==2.27.1
rsa==4.8
six==1.16.0
spotipy==2.19.0
Telethon==1.24.0
urllib3==1.26.9
